Move test serializers from TestObjects.cs to new file TestSerializers.cs
